separated test cases into 2 groups.
moved the larger brla test case into the tests/ directory instead of
src/r2b2/tests as it is time consuming. Edited the tox.ini file to
create a py37full environment which runs tests from both directories and
modified py37 to only run tests within the src/r2b2/tests.
