+++ src/r2b2/tests/test_brla.py
@@ -1,10 +1,7 @@
 import pytest
-import json
 
 import r2b2.tests.util as util
 from r2b2.brla import BayesianRLA as BRLA
-from r2b2.contest import Contest
-from r2b2.contest import ContestType
 
 default_contest = util.generate_contest(100)
 
@@ -42,17 +39,3 @@ def test_exceptions():
         simplebrla.compute_all_min_winner_ballots(0)
     with pytest.raises(ValueError):
         simplebrla.compute_all_min_winner_ballots(1000)
-
-
-def test_paper_data():
-    """Test brla using data from BRLA paper Table 3."""
-    with open('src/r2b2/tests/brla_test_data.json', 'r') as json_file:
-        data = json.load(json_file)
-
-    rounds = data['rounds']
-    contest = Contest(100000, {'A':60000, 'B':40000}, 1, ['A'], ContestType.PLURALITY)
-    
-    for test in data['test_cases']:
-        audit = BRLA(data['test_cases'][test]['risk'], 1.0, contest)
-        kmins = audit.compute_min_winner_ballots(rounds)
-        assert kmins == data['test_cases'][test]['kmins']
