+++ tox.ini
@@ -21,7 +21,7 @@ ignore_basepython_conflict = true
 
 [testenv]
 basepython =
-    {py37,docs}: {env:TOXPYTHON:python3.7}
+    {py37,docs,py37full}: {env:TOXPYTHON:python3.7}
     {bootstrap,clean,check,report,codecov}: {env:TOXPYTHON:python3}
 setenv =
     PYTHONPATH={toxinidir}/tests
@@ -35,7 +35,7 @@ deps =
     pytest-cov
     -r{toxinidir}/requirements.txt
 commands =
-    {posargs:pytest --cov --cov-report=term-missing -vv }
+    {posargs:pytest src --cov --cov-report=term-missing -vv }
 
 [testenv:check]
 deps =
@@ -79,3 +79,12 @@ commands =
 commands = coverage erase
 skip_install = true
 deps = coverage
+
+[testenv:py37full]
+deps =
+    pytest
+    pytest-travis-fold
+    pytest-cov
+    -r{toxinidir}/requirements.txt
+commands =
+    {posargs:pytest --cov --cov-report=term-missing -vv }
